requirement.txt
absl-py==2.0.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.2.2
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.9.0
fonttools==4.44.3
fqdn==1.5.1
fsspec==2023.10.0
ftfy==6.1.3
fvcore==0.1.5.post20221221
gmpy2==2.1.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grad-cam==1.5.0
grpcio==1.59.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.4
imageio==2.32.0
importlib_metadata==8.5.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lazy_loader==0.3
loguru==0.7.2
Markdown==3.5.1
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.1
notebook_shim==0.2.4
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.7.0
packaging==24.1
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3
platformdirs==4.3.6
portalocker==3.0.0
prettytable==3.12.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.18.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.8.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.3
setuptools==68.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.10
tinycss2==1.3.0
tokenizers==0.21.0
tome==0.1
torch==2.1.1
torchaudio==2.1.1
torchcam==0.4.0
torchvision==0.16.1
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.49.0
triton==2.1.0
ttach==0.0.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.12
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
wheel==0.41.2
wordcloud==1.9.4
yacs==0.1.8
zipp==3.20.2
